pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k-daytrade-morning-agent"
version = "0.1.0"
description = "Automated pre-market Korean day-trading research brief generator."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "k-daytrade-morning-agent maintainers" }]
dependencies = [
  "Jinja2>=3.1.4",
  "openai>=1.93.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.6.0",
]

[project.scripts]
daytrade-agent = "daytrade_agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["daytrade_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["dist", "private_artifacts", ".codex-report-workspace", ".runtime"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
